Lookup Limits

Looks up the long or short-term water quality limits for BC. If the limits depend on on the pH, total hardness (CaCO3), total chloride or the concentration of methyl mercury and site specific values are not provided then the dependent limits are returned as missing values.

lookup_limits(
  ph = NULL,
  hardness = NULL,
  chloride = NULL,
  methyl_mercury = NULL,
  term = "long",
  use = "Freshwater Life"
)

Arguments

ph
A number indicating the pH in pH units at the site of interest.

hardness
A number indicating the total hardness (CaCO3) in mg/L at the site of interest.

chloride
A number indicating the total chloride concentration in mg/L at the site of interest.

methyl_mercury
A number indicating the total concentration of methyl mercury in ug/L at the site of interest.

term
A string indicating whether to lookup the "long" or "short"-term limits.

use
A string indicating the Use.

See also

Examples

lookup_limits(ph = 8, hardness = 100, chloride = 50, methyl_mercury = 2)

#>                                                                              Variable
#> 1                                     1,1,2-Trichloroethene (Trichloroethylene) Total
#> 2             1,1,2,2-Tetrachloroethene (Tetrachloroethylene, Perchlorethylene) Total
#> 3                                                           1,2-Dichlorobenzene Total
#> 4                                                            1,2-Dichloroethane Total
#> 5                                                          1,2-Propylene glycol Total
#> 6                                                        1,2,3-Trichlorobenzene Total
#> 7                                                    1,2,3,4-Tetrachlorobenzene Total
#> 8                                                        1,2,4-Trichlorobenzene Total
#> 9                                                           1,3-Dichlorobenzene Total
#> 10                                                          1,4-Dichlorobenzene Total
#> 11                                       3-Lodo-2-propynl butylcarbamate (IPBC) Total
#> 12                                                                 Acenaphthene Total
#> 13                                                                     Aldicarb Total
#> 14                                                          Alkalinity ( CaCO3) Total
#> 15                                                                 Aluminum Dissolved
#> 16                                                                      Ammonia Total
#> 17                                                               Antimony (III) Total
#> 18                                                                       Barium Total
#> 19                                                                      Benzene Total
#> 20                                                               Benzo(a)pyrene Total
#> 21                                                                    Beryllium Total
#> 22                                                                        Boron Total
#> 23                                                                     Bromacil Total
#> 24                                                                   Bromoxynil Total
#> 25                                                                  Cadmium Dissolved
#> 26                                                                       Captan Total
#> 27                                                                   Carbofuran Total
#> 28                                                                 Chlorate Dissolved
#> 29                                                                 Chloride Dissolved
#> 30                                                            Chlorine Residual Total
#> 31        Chlorothalonil (2,4,5,6-Tetrachloro-1,3-benzenecarbonitrile, Daconil) Total
#> 32                                                                 Chlorpyrifos Total
#> 33                                                          Chromium hexavalent Total
#> 34                                                           Chromium trivalent Total
#> 35                                                                       Cobalt Total
#> 36                                                                    Cyanazine Total
#> 37                                                Cyanide weak acid dissociable Total
#> 38                                                                 Deltamethrin Total
#> 39                                           Di-(2-ethylhexyl) phthalate (DEHP) Total
#> 40                                                               Di-n-butyl tin Total
#> 41                                 Dibutyl phthalate (DBP, di-n-butylphthalate) Total
#> 42                                                                      Dicamba Total
#> 43                                         Dichloromethane (Methylene chloride) Total
#> 44                                           Dichlorophenoxyacetic acid (2,4-D) Total
#> 45                                                              Diclofop-methyl Total
#> 46                                    Didecyl dimethyl ammonium chloride (DDAC) Total
#> 47                                                                   Dimethoate Total
#> 48                                                                      Dinoseb Total
#> 49                                                         Ethinylestradiol 17a Total
#> 50                                                                 Ethylbenzene Total
#> 51                                                              Ethylene glycol Total
#> 52                                                                     Fluorene Total
#> 53                                                          Hexachlorobutadiene Total
#> 54                                              Hexachlorocyclohexane (Lindane) Total
#> 55                                                                         Lead Total
#> 56                                                                      Linuron Total
#> 57                                                                    Malathion Total
#> 58                                                                    Manganese Total
#> 59                                                                      Mercury Total
#> 60 Methylchlorophenoxy acetic acid (4-chloro-2-methylphenoxy acetic acid; MCPA) Total
#> 61                                                                  Metolachlor Total
#> 62                                                                   Metribuzin Total
#> 63                                                                   Molybdenum Total
#> 64                                                            Monochlorobenzene Total
#> 65                                                                  Naphthalene Total
#> 66                                                                       Nickel Total
#> 67                                                                      Nitrate Total
#> 68                                                                      Nitrite Total
#> 69                                                           Pentachlorobenzene Total
#> 70                                                                 Phenanthrene Total
#> 71                                                                     Picloram Total
#> 72                                                                    Quinoline Total
#> 73                                                                     Selenium Total
#> 74                                                                       Silver Total
#> 75                                                                     Simazine Total
#> 76                                                                      Styrene Total
#> 77                                                                     Sulphate Total
#> 78                                    Tetrachloromethane (carbon tetrachloride) Total
#> 79                                                                     Thallium Total
#> 80                                                             Tin, Tributyltin Total
#> 81                                                                      Toluene Total
#> 82                                                                    Triallate Total
#> 83                                                Trichloromethane (chloroform) Total
#> 84                                                                  Trifluralin Total
#> 85                                                                      Uranium Total
#> 86                                                                       Xylene Total
#> 87                                                                         Zinc Total
#>     UpperLimit Units
#> 1   21.0000000  ug/L
#> 2  110.0000000  ug/L
#> 3    0.7000000  ug/L
#> 4  100.0000000  ug/L
#> 5  500.0000000  mg/L
#> 6    8.0000000  ug/L
#> 7    1.8000000  ug/L
#> 8   24.0000000  ug/L
#> 9  150.0000000  ug/L
#> 10  26.0000000  ug/L
#> 11   1.9000000  ug/L
#> 12   6.0000000  ug/L
#> 13   1.0000000  ug/L
#> 14          NA  mg/L
#> 15   0.0500000  mg/L
#> 16   1.2412358  mg/L
#> 17   9.0000000  ug/L
#> 18   1.0000000  mg/L
#> 19  40.0000000  ug/L
#> 20   0.0100000  ug/L
#> 21   0.1300000  ug/L
#> 22   1.2000000  mg/L
#> 23   5.0000000  ug/L
#> 24   5.0000000  ug/L
#> 25   0.2114864  ug/L
#> 26   1.3000000  ug/L
#> 27   1.8000000  ug/L
#> 28  30.0000000  mg/L
#> 29 150.0000000  mg/L
#> 30   2.0000000  ug/L
#> 31   0.1800000  ug/L
#> 32   0.0020000  ug/L
#> 33   1.0000000  ug/L
#> 34   8.9000000  ug/L
#> 35   4.0000000  ug/L
#> 36   2.0000000  ug/L
#> 37   5.0000000  ug/L
#> 38   0.0004000  ug/L
#> 39  16.0000000  ug/L
#> 40   0.0800000  ug/L
#> 41  19.0000000  ug/L
#> 42  10.0000000  ug/L
#> 43  98.1000000  ug/L
#> 44   4.0000000  ug/L
#> 45   6.1000000  ug/L
#> 46   1.5000000  ug/L
#> 47   6.2000000  ug/L
#> 48   0.0500000  ug/L
#> 49   0.5000000  ng/L
#> 50   0.2000000  mg/L
#> 51 192.0000000  mg/L
#> 52  12.0000000  ug/L
#> 53   1.3000000  ug/L
#> 54   0.0100000  ug/L
#> 55   6.4947750  ug/L
#> 56   7.0000000  ug/L
#> 57   0.1000000  ug/L
#> 58   1.0450000  mg/L
#> 59   0.0000500  ug/L
#> 60   2.6000000  ug/L
#> 61   7.8000000  ug/L
#> 62   1.0000000  ug/L
#> 63   1.0000000  mg/L
#> 64   1.3000000  ug/L
#> 65   1.0000000  ug/L
#> 66  95.5767263  ug/L
#> 67   3.0000000  mg/L
#> 68   0.2000000  mg/L
#> 69   6.0000000  ug/L
#> 70   0.3000000  ug/L
#> 71  29.0000000  ug/L
#> 72   3.4000000  ug/L
#> 73   2.0000000  ug/L
#> 74   0.0500000  ug/L
#> 75  10.0000000  ug/L
#> 76  72.0000000  ug/L
#> 77 309.0000000  mg/L
#> 78  13.3000000  ug/L
#> 79   0.8000000  ug/L
#> 80   0.0080000  ug/L
#> 81   0.5000000  ug/L
#> 82   0.2400000  ug/L
#> 83   1.8000000  ug/L
#> 84   0.2000000  ug/L
#> 85   8.5000000  ug/L
#> 86   0.0300000  mg/L
#> 87  15.0000000  ug/L

lookup_limits(term = "short")

#>                               Variable UpperLimit Units
#> 1                   Aluminum Dissolved         NA  mg/L
#> 2                        Ammonia Total         NA  mg/L
#> 3                        Arsenic Total       5.00  ug/L
#> 4                    Cadmium Dissolved         NA  ug/L
#> 5                   Chloride Dissolved     600.00  mg/L
#> 6          Chlorophenol, 2,3-DCP Total         NA  ug/L
#> 7        Chlorophenol, 2,3,4-TCP Total         NA  ug/L
#> 8     Chlorophenol, 2,3,4,5-TTCP Total         NA  ug/L
#> 9    Chlorophenol, 2,3,4,5,6-PCP Total         NA  ug/L
#> 10    Chlorophenol, 2,3,4,6-TTCP Total         NA  ug/L
#> 11       Chlorophenol, 2,3,5-TCP Total         NA  ug/L
#> 12    Chlorophenol, 2,3,5,6-TTCP Total         NA  ug/L
#> 13       Chlorophenol, 2,3,6-TCP Total         NA  ug/L
#> 14         Chlorophenol, 2,4-DCP Total         NA  ug/L
#> 15       Chlorophenol, 2,4,5-TCP Total         NA  ug/L
#> 16       Chlorophenol, 2,4,6-TCP Total         NA  ug/L
#> 17         Chlorophenol, 2,5-DCP Total         NA  ug/L
#> 18         Chlorophenol, 2,6-DCP Total         NA  ug/L
#> 19         Chlorophenol, 3,4-DCP Total         NA  ug/L
#> 20       Chlorophenol, 3,4,5-TCP Total         NA  ug/L
#> 21         Chlorophenol, 3,5-DCP Total         NA  ug/L
#> 22                        Cobalt Total     110.00  ug/L
#> 23 Cyanide weak acid dissociable Total      10.00  ug/L
#> 24     Dehydroabietic acid (DHA) Total         NA  ug/L
#> 25          Ethinylestradiol 17a Total       0.75  ng/L
#> 26                      Fluoride Total         NA  mg/L
#> 27                      Iron Dissolved       0.35  mg/L
#> 28                          Iron Total       1.00  mg/L
#> 29                          Lead Total         NA  ug/L
#> 30                     Manganese Total         NA  mg/L
#> 31   Methyl tertiary-butyl ether Total       3.40  mg/L
#> 32                    Molybdenum Total       2.00  mg/L
#> 33                   Naphthalene Total       1.00  ug/L
#> 34                       Nitrate Total      32.80  mg/L
#> 35                       Nitrite Total         NA  mg/L
#> 36     Polychlorinated Biphenyls Total       0.10  ng/L
#> 37                        Silver Total         NA  ug/L
#> 38                          Zinc Total         NA  ug/L