Lookup Limits

Looks up the long or short-term water quality limits for BC. If the limits depend on on the pH, total hardness (CaCO3), total chloride or the concentration of methyl mercury and site specific values are not provided then the dependent limits are returned as missing values.

lookup_limits(
  ph = NULL,
  hardness = NULL,
  chloride = NULL,
  methyl_mercury = NULL,
  term = "long",
  use = "Freshwater Life"
)

Arguments

ph
A number indicating the pH in pH units at the site of interest.

hardness
A number indicating the total hardness (CaCO3) in mg/L at the site of interest.

chloride
A number indicating the total chloride concentration in mg/L at the site of interest.

methyl_mercury
A number indicating the total concentration of methyl mercury in ug/L at the site of interest.

term
A string indicating whether to lookup the "long" or "short"-term limits.

use
A string indicating the Use.

See also

Examples

lookup_limits(ph = 8, hardness = 100, chloride = 50, methyl_mercury = 2)

#> # A tibble: 87 x 3
#>    Variable                                                     UpperLimit Units
#>    <fct>                                                             <dbl> <fct>
#>  1 1,1,2-Trichloroethene (Trichloroethylene) Total                    21   ug/L 
#>  2 1,1,2,2-Tetrachloroethene (Tetrachloroethylene, Perchloreth…      110   ug/L 
#>  3 1,2-Dichlorobenzene Total                                           0.7 ug/L 
#>  4 1,2-Dichloroethane Total                                          100   ug/L 
#>  5 1,2-Propylene glycol Total                                        500   mg/L 
#>  6 1,2,3-Trichlorobenzene Total                                        8   ug/L 
#>  7 1,2,3,4-Tetrachlorobenzene Total                                    1.8 ug/L 
#>  8 1,2,4-Trichlorobenzene Total                                       24   ug/L 
#>  9 1,3-Dichlorobenzene Total                                         150   ug/L 
#> 10 1,4-Dichlorobenzene Total                                          26   ug/L 
#> # … with 77 more rows

lookup_limits(term = "short")

#> # A tibble: 38 x 3
#>    Variable                          UpperLimit Units
#>    <fct>                                  <dbl> <fct>
#>  1 Aluminum Dissolved                        NA mg/L 
#>  2 Ammonia Total                             NA mg/L 
#>  3 Arsenic Total                              5 ug/L 
#>  4 Cadmium Dissolved                         NA ug/L 
#>  5 Chloride Dissolved                       600 mg/L 
#>  6 Chlorophenol, 2,3-DCP Total               NA ug/L 
#>  7 Chlorophenol, 2,3,4-TCP Total             NA ug/L 
#>  8 Chlorophenol, 2,3,4,5-TTCP Total          NA ug/L 
#>  9 Chlorophenol, 2,3,4,5,6-PCP Total         NA ug/L 
#> 10 Chlorophenol, 2,3,4,6-TTCP Total          NA ug/L 
#> # … with 28 more rows