Lookup Variables

Returns recognised water quality variables. If codes = NULL the function returns all recognised variable names. Otherwise it looks up the matching variables from codes. Whether or not the codes are compressed or expanded is unimportant.

lookup_variables(
  codes = NULL,
  messages = getOption("wqbc.messages", default = TRUE)
)

Arguments

codes
An optional character vector of codes to look up variables.

messages
A flag indicating whether to print messages.

See also

Examples

lookup_variables()

#>   [1] "1,1-Dichloroethylene"                                                              
#>   [2] "1,1,2-Trichloroethene (Trichloroethylene) Total"                                   
#>   [3] "1,1,2,2-Tetrachloroethene (Tetrachloroethylene, Perchlorethylene) Total"           
#>   [4] "1,2-Dichlorobenzene Total"                                                         
#>   [5] "1,2-Dichloroethane Total"                                                          
#>   [6] "1,2-Propylene glycol Total"                                                        
#>   [7] "1,2,3-Trichlorobenzene Total"                                                      
#>   [8] "1,2,3,4-Tetrachlorobenzene Total"                                                  
#>   [9] "1,2,4-Trichlorobenzene Total"                                                      
#>  [10] "1,3-Dichlorobenzene Total"                                                         
#>  [11] "1,4-Dichlorobenzene Total"                                                         
#>  [12] "3-Lodo-2-propynl butylcarbamate (IPBC) Total"                                      
#>  [13] "Acenaphthene Total"                                                                
#>  [14] "Aldicarb Total"                                                                    
#>  [15] "Alkalinity ( CaCO3) Total"                                                         
#>  [16] "Alkalinity pH 4.5/4.2"                                                             
#>  [17] "Alkalinity Total"                                                                  
#>  [18] "Aluminum Dissolved"                                                                
#>  [19] "Aluminum Total"                                                                    
#>  [20] "Ammonia Dissolved"                                                                 
#>  [21] "Ammonia Total"                                                                     
#>  [22] "Antimony (III) Total"                                                              
#>  [23] "Arsenic Dissolved"                                                                 
#>  [24] "Arsenic Total"                                                                     
#>  [25] "Atrazine"                                                                          
#>  [26] "Barium Dissolved"                                                                  
#>  [27] "Barium Total"                                                                      
#>  [28] "Benzene Total"                                                                     
#>  [29] "Benzo(a)pyrene Total"                                                              
#>  [30] "Beryllium Total"                                                                   
#>  [31] "Biochemical Oxygen Demand"                                                         
#>  [32] "Boron Dissolved"                                                                   
#>  [33] "Boron Total"                                                                       
#>  [34] "Bromacil Total"                                                                    
#>  [35] "Bromoxynil Total"                                                                  
#>  [36] "Cadmium Dissolved"                                                                 
#>  [37] "Cadmium Total"                                                                     
#>  [38] "Calcium Dissolved"                                                                 
#>  [39] "Captan Total"                                                                      
#>  [40] "Carbofuran Total"                                                                  
#>  [41] "Chlorate Dissolved"                                                                
#>  [42] "Chloride Dissolved"                                                                
#>  [43] "Chloride Total"                                                                    
#>  [44] "Chlorine Residual Total"                                                           
#>  [45] "Chlorite Dissolved"                                                                
#>  [46] "Chlorophenol, 2,3-DCP Total"                                                       
#>  [47] "Chlorophenol, 2,3,4-TCP Total"                                                     
#>  [48] "Chlorophenol, 2,3,4,5-TTCP Total"                                                  
#>  [49] "Chlorophenol, 2,3,4,5,6-PCP Total"                                                 
#>  [50] "Chlorophenol, 2,3,4,6-TTCP Total"                                                  
#>  [51] "Chlorophenol, 2,3,5-TCP Total"                                                     
#>  [52] "Chlorophenol, 2,3,5,6-TTCP Total"                                                  
#>  [53] "Chlorophenol, 2,3,6-TCP Total"                                                     
#>  [54] "Chlorophenol, 2,4-DCP Total"                                                       
#>  [55] "Chlorophenol, 2,4,5-TCP Total"                                                     
#>  [56] "Chlorophenol, 2,4,6-TCP Total"                                                     
#>  [57] "Chlorophenol, 2,5-DCP Total"                                                       
#>  [58] "Chlorophenol, 2,6-DCP Total"                                                       
#>  [59] "Chlorophenol, 3,4-DCP Total"                                                       
#>  [60] "Chlorophenol, 3,4,5-TCP Total"                                                     
#>  [61] "Chlorophenol, 3,5-DCP Total"                                                       
#>  [62] "Chlorophenol, DCP"                                                                 
#>  [63] "Chlorophenol, TCP"                                                                 
#>  [64] "Chlorophenol, TTCP"                                                                
#>  [65] "Chlorophyll A"                                                                     
#>  [66] "Chlorothalonil (2,4,5,6-Tetrachloro-1,3-benzenecarbonitrile, Daconil) Total"       
#>  [67] "Chlorpyrifos Total"                                                                
#>  [68] "Chromium Dissolved"                                                                
#>  [69] "Chromium hexavalent Total"                                                         
#>  [70] "Chromium Total"                                                                    
#>  [71] "Chromium trivalent Total"                                                          
#>  [72] "Chrysene"                                                                          
#>  [73] "Cobalt Dissolved"                                                                  
#>  [74] "Cobalt Total"                                                                      
#>  [75] "Coliforms Fecal"                                                                   
#>  [76] "Colour Apparent"                                                                   
#>  [77] "Colour True"                                                                       
#>  [78] "Conductivity Specific"                                                             
#>  [79] "Conductivity Specific - Field"                                                     
#>  [80] "Copper Dissolved"                                                                  
#>  [81] "Copper Total"                                                                      
#>  [82] "Cyanazine Total"                                                                   
#>  [83] "Cyanide Dissolved"                                                                 
#>  [84] "Cyanide Total"                                                                     
#>  [85] "Cyanide weak acid dissociable Total"                                               
#>  [86] "Dehydroabietic acid (DHA) Total"                                                   
#>  [87] "Deltamethrin Total"                                                                
#>  [88] "Di-(2-ethylhexyl) phthalate (DEHP) Total"                                          
#>  [89] "Di-n-butyl tin Total"                                                              
#>  [90] "Dibutyl phthalate (DBP, di-n-butylphthalate) Total"                                
#>  [91] "Dicamba Total"                                                                     
#>  [92] "Dichlorobromomethane (Bromodichloromethane)"                                       
#>  [93] "Dichloromethane (Methylene chloride) Total"                                        
#>  [94] "Dichlorophenoxyacetic acid (2,4-D) Total"                                          
#>  [95] "Diclofop-methyl Total"                                                             
#>  [96] "Didecyl dimethyl ammonium chloride (DDAC) Total"                                   
#>  [97] "Dimethoate Total"                                                                  
#>  [98] "Dinoseb Total"                                                                     
#>  [99] "Dissolved Solids (Filterable Residue)"                                             
#> [100] "Enterococci"                                                                       
#> [101] "Escherichia coli"                                                                  
#> [102] "Ethinylestradiol 17a Total"                                                        
#> [103] "Ethylbenzene Total"                                                                
#> [104] "Ethylene glycol Total"                                                             
#> [105] "Extinction Depth"                                                                  
#> [106] "Fluorene Total"                                                                    
#> [107] "Fluoride Dissolved"                                                                
#> [108] "Fluoride Total"                                                                    
#> [109] "Hardness Dissolved"                                                                
#> [110] "Hardness Total"                                                                    
#> [111] "Hexachlorobenzene"                                                                 
#> [112] "Hexachlorobutadiene Total"                                                         
#> [113] "Hexachlorocyclohexane (Lindane) Total"                                             
#> [114] "Iron Dissolved"                                                                    
#> [115] "Iron Total"                                                                        
#> [116] "Lead Dissolved"                                                                    
#> [117] "Lead Total"                                                                        
#> [118] "Linuron Total"                                                                     
#> [119] "Lithium Total"                                                                     
#> [120] "Malathion Total"                                                                   
#> [121] "Manganese Dissolved"                                                               
#> [122] "Manganese Total"                                                                   
#> [123] "Mercury Dissolved"                                                                 
#> [124] "Mercury Methyl"                                                                    
#> [125] "Mercury Total"                                                                     
#> [126] "Methyl tertiary-butyl ether Total"                                                 
#> [127] "Methylchlorophenoxy acetic acid (4-chloro-2-methylphenoxy acetic acid; MCPA) Total"
#> [128] "Metolachlor Total"                                                                 
#> [129] "Metribuzin Total"                                                                  
#> [130] "Molybdenum Dissolved"                                                              
#> [131] "Molybdenum Total"                                                                  
#> [132] "Monochlorobenzene Total"                                                           
#> [133] "Naphthalene D8"                                                                    
#> [134] "Naphthalene Total"                                                                 
#> [135] "Nickel Dissolved"                                                                  
#> [136] "Nickel Total"                                                                      
#> [137] "Nitrate Nitrite"                                                                   
#> [138] "Nitrate Nitrite Dissolved"                                                         
#> [139] "Nitrate Soluble"                                                                   
#> [140] "Nitrate Total"                                                                     
#> [141] "Nitrite Total"                                                                     
#> [142] "Nitrogen Dissolved"                                                                
#> [143] "Nitrogen Total"                                                                    
#> [144] "Organic Carbon Dissolved"                                                          
#> [145] "Organic Carbon Total"                                                              
#> [146] "Oxygen Dissolved"                                                                  
#> [147] "Oxygen Dissolved Field"                                                            
#> [148] "Pentachlorobenzene Total"                                                          
#> [149] "pH"                                                                                
#> [150] "Phenanthrene Total"                                                                
#> [151] "Phosphorus Dissolved"                                                              
#> [152] "Phosphorus Ortho"                                                                  
#> [153] "Phosphorus Total"                                                                  
#> [154] "Picloram Total"                                                                    
#> [155] "Polychlorinated Biphenyls (PCB) Total"                                             
#> [156] "Polycyclic Aromatic Hydrocarbons (Carcinogenic) Total"                             
#> [157] "Pseudomonas Aeruginosa"                                                            
#> [158] "Quinoline Total"                                                                   
#> [159] "Salinity"                                                                          
#> [160] "Selenium Dissolved"                                                                
#> [161] "Selenium Total"                                                                    
#> [162] "Silver Dissolved"                                                                  
#> [163] "Silver Total"                                                                      
#> [164] "Simazine Total"                                                                    
#> [165] "Strontium Total"                                                                   
#> [166] "Styrene Total"                                                                     
#> [167] "Sulphate Dissolved"                                                                
#> [168] "Sulphate Total"                                                                    
#> [169] "Temperature"                                                                       
#> [170] "Temperature - Field"                                                               
#> [171] "Tetrachloromethane (carbon tetrachloride) Total"                                   
#> [172] "Thallium Dissolved"                                                                
#> [173] "Thallium Total"                                                                    
#> [174] "Tin, Tributyltin Total"                                                            
#> [175] "Toluene Total"                                                                     
#> [176] "Triallate Total"                                                                   
#> [177] "Tribromomethane (Bromoform)"                                                       
#> [178] "Trichloromethane (chloroform) Total"                                               
#> [179] "Trifluralin Total"                                                                 
#> [180] "Turbidity"                                                                         
#> [181] "Turbidity - Field"                                                                 
#> [182] "Uranium Dissolved"                                                                 
#> [183] "Uranium Total"                                                                     
#> [184] "Vanadium Total"                                                                    
#> [185] "Vinyl Chloride"                                                                    
#> [186] "Xylene Total"                                                                      
#> [187] "Zinc Dissolved"                                                                    
#> [188] "Zinc Total"                                                                        

lookup_variables(c("AL-D", "EMS_AS_T", "B--T", "KRYP"), messages = TRUE)

#> Replaced 'AL-D' with 'Aluminum Dissolved', 'AS-T' with 'Arsenic Total' and 'B--T' with 'Boron Total'.

#> Failed to replace 'KRYP'.

#> [1] "Aluminum Dissolved" "Arsenic Total"      "Boron Total"       
#> [4] NA