Lookup Variables

Returns recognised water quality variables. If codes = NULL the function returns all recognised variable names. Otherwise it looks up the matching variables from codes. Whether or not the codes are compressed or expanded is unimportant.

lookup_variables(
  codes = NULL,
  messages = getOption("wqbc.messages", default = TRUE)
)

Arguments

codes
An optional character vector of codes to look up variables.

messages
A flag indicating whether to print messages.

See also

Examples

lookup_variables()

#>   [1] "1,1,2-Trichloroethene (Trichloroethylene) Total"                                   
#>   [2] "1,1,2,2-Tetrachloroethene (Tetrachloroethylene, Perchlorethylene) Total"           
#>   [3] "1,2-Dichlorobenzene Total"                                                         
#>   [4] "1,2-Dichloroethane Total"                                                          
#>   [5] "1,2-Propylene glycol Total"                                                        
#>   [6] "1,2,3-Trichlorobenzene Total"                                                      
#>   [7] "1,2,3,4-Tetrachlorobenzene Total"                                                  
#>   [8] "1,2,4-Trichlorobenzene Total"                                                      
#>   [9] "1,3-Dichlorobenzene Total"                                                         
#>  [10] "1,4-Dichlorobenzene Total"                                                         
#>  [11] "3-Lodo-2-propynl butylcarbamate (IPBC) Total"                                      
#>  [12] "Acenaphthene Total"                                                                
#>  [13] "Aldicarb Total"                                                                    
#>  [14] "Alkalinity ( CaCO3) Total"                                                         
#>  [15] "Alkalinity pH 4.5/4.2"                                                             
#>  [16] "Alkalinity Total"                                                                  
#>  [17] "Aluminum Dissolved"                                                                
#>  [18] "Ammonia Dissolved"                                                                 
#>  [19] "Ammonia Total"                                                                     
#>  [20] "Antimony (III) Total"                                                              
#>  [21] "Arsenic Dissolved"                                                                 
#>  [22] "Arsenic Total"                                                                     
#>  [23] "Barium Dissolved"                                                                  
#>  [24] "Barium Total"                                                                      
#>  [25] "Benzene Total"                                                                     
#>  [26] "Benzo(a)pyrene Total"                                                              
#>  [27] "Beryllium Total"                                                                   
#>  [28] "Boron Dissolved"                                                                   
#>  [29] "Boron Total"                                                                       
#>  [30] "Bromacil Total"                                                                    
#>  [31] "Bromoxynil Total"                                                                  
#>  [32] "Cadmium Dissolved"                                                                 
#>  [33] "Cadmium Total"                                                                     
#>  [34] "Captan Total"                                                                      
#>  [35] "Carbofuran Total"                                                                  
#>  [36] "Chlorate Dissolved"                                                                
#>  [37] "Chloride Dissolved"                                                                
#>  [38] "Chloride Total"                                                                    
#>  [39] "Chlorine Residual Total"                                                           
#>  [40] "Chlorophenol, 2,3-DCP Total"                                                       
#>  [41] "Chlorophenol, 2,3,4-TCP Total"                                                     
#>  [42] "Chlorophenol, 2,3,4,5-TTCP Total"                                                  
#>  [43] "Chlorophenol, 2,3,4,5,6-PCP Total"                                                 
#>  [44] "Chlorophenol, 2,3,4,6-TTCP Total"                                                  
#>  [45] "Chlorophenol, 2,3,5-TCP Total"                                                     
#>  [46] "Chlorophenol, 2,3,5,6-TTCP Total"                                                  
#>  [47] "Chlorophenol, 2,3,6-TCP Total"                                                     
#>  [48] "Chlorophenol, 2,4-DCP Total"                                                       
#>  [49] "Chlorophenol, 2,4,5-TCP Total"                                                     
#>  [50] "Chlorophenol, 2,4,6-TCP Total"                                                     
#>  [51] "Chlorophenol, 2,5-DCP Total"                                                       
#>  [52] "Chlorophenol, 2,6-DCP Total"                                                       
#>  [53] "Chlorophenol, 3,4-DCP Total"                                                       
#>  [54] "Chlorophenol, 3,4,5-TCP Total"                                                     
#>  [55] "Chlorophenol, 3,5-DCP Total"                                                       
#>  [56] "Chlorothalonil (2,4,5,6-Tetrachloro-1,3-benzenecarbonitrile, Daconil) Total"       
#>  [57] "Chlorpyrifos Total"                                                                
#>  [58] "Chromium Dissolved"                                                                
#>  [59] "Chromium hexavalent Total"                                                         
#>  [60] "Chromium Total"                                                                    
#>  [61] "Chromium trivalent Total"                                                          
#>  [62] "Cobalt Dissolved"                                                                  
#>  [63] "Cobalt Total"                                                                      
#>  [64] "Coliforms Fecal"                                                                   
#>  [65] "Colour Apparent"                                                                   
#>  [66] "Colour True"                                                                       
#>  [67] "Copper Dissolved"                                                                  
#>  [68] "Copper Total"                                                                      
#>  [69] "Cyanazine Total"                                                                   
#>  [70] "Cyanide Dissolved"                                                                 
#>  [71] "Cyanide Total"                                                                     
#>  [72] "Cyanide weak acid dissociable Total"                                               
#>  [73] "Dehydroabietic acid (DHA) Total"                                                   
#>  [74] "Deltamethrin Total"                                                                
#>  [75] "Di-(2-ethylhexyl) phthalate (DEHP) Total"                                          
#>  [76] "Di-n-butyl tin Total"                                                              
#>  [77] "Dibutyl phthalate (DBP, di-n-butylphthalate) Total"                                
#>  [78] "Dicamba Total"                                                                     
#>  [79] "Dichloromethane (Methylene chloride) Total"                                        
#>  [80] "Dichlorophenoxyacetic acid (2,4-D) Total"                                          
#>  [81] "Diclofop-methyl Total"                                                             
#>  [82] "Didecyl dimethyl ammonium chloride (DDAC) Total"                                   
#>  [83] "Dimethoate Total"                                                                  
#>  [84] "Dinoseb Total"                                                                     
#>  [85] "E. coli"                                                                           
#>  [86] "Ethinylestradiol 17a Total"                                                        
#>  [87] "Ethylbenzene Total"                                                                
#>  [88] "Ethylene glycol Total"                                                             
#>  [89] "Fluorene Total"                                                                    
#>  [90] "Fluoride Dissolved"                                                                
#>  [91] "Fluoride Total"                                                                    
#>  [92] "Hardness Dissolved"                                                                
#>  [93] "Hardness Total"                                                                    
#>  [94] "Hexachlorobutadiene Total"                                                         
#>  [95] "Hexachlorocyclohexane (Lindane) Total"                                             
#>  [96] "Iron Dissolved"                                                                    
#>  [97] "Iron Total"                                                                        
#>  [98] "Lead Dissolved"                                                                    
#>  [99] "Lead Total"                                                                        
#> [100] "Linuron Total"                                                                     
#> [101] "Malathion Total"                                                                   
#> [102] "Manganese Dissolved"                                                               
#> [103] "Manganese Total"                                                                   
#> [104] "Mercury Dissolved"                                                                 
#> [105] "Mercury Methyl"                                                                    
#> [106] "Mercury Total"                                                                     
#> [107] "Methyl tertiary-butyl ether Total"                                                 
#> [108] "Methylchlorophenoxy acetic acid (4-chloro-2-methylphenoxy acetic acid; MCPA) Total"
#> [109] "Metolachlor Total"                                                                 
#> [110] "Metribuzin Total"                                                                  
#> [111] "Molybdenum Dissolved"                                                              
#> [112] "Molybdenum Total"                                                                  
#> [113] "Monochlorobenzene Total"                                                           
#> [114] "Naphthalene D8"                                                                    
#> [115] "Naphthalene Total"                                                                 
#> [116] "Nickel Dissolved"                                                                  
#> [117] "Nickel Total"                                                                      
#> [118] "Nitrate Dissolved"                                                                 
#> [119] "Nitrate Nitrite"                                                                   
#> [120] "Nitrate Soluble"                                                                   
#> [121] "Nitrate Total"                                                                     
#> [122] "Nitrite Total"                                                                     
#> [123] "Nitrogen Dissolved"                                                                
#> [124] "Nitrogen Total"                                                                    
#> [125] "Oxygen Dissolved"                                                                  
#> [126] "Pentachlorobenzene Total"                                                          
#> [127] "pH"                                                                                
#> [128] "Phenanthrene Total"                                                                
#> [129] "Phosphorus Dissolved"                                                              
#> [130] "Phosphorus Total"                                                                  
#> [131] "Picloram Total"                                                                    
#> [132] "Polychlorinated Biphenyls Total"                                                   
#> [133] "Quinoline Total"                                                                   
#> [134] "Selenium Dissolved"                                                                
#> [135] "Selenium Total"                                                                    
#> [136] "Silver Dissolved"                                                                  
#> [137] "Silver Total"                                                                      
#> [138] "Simazine Total"                                                                    
#> [139] "Styrene Total"                                                                     
#> [140] "Sulphate Dissolved"                                                                
#> [141] "Sulphate Total"                                                                    
#> [142] "Temperature"                                                                       
#> [143] "Tetrachloromethane (carbon tetrachloride) Total"                                   
#> [144] "Thallium Dissolved"                                                                
#> [145] "Thallium Total"                                                                    
#> [146] "Tin, Tributyltin Total"                                                            
#> [147] "Toluene Total"                                                                     
#> [148] "Triallate Total"                                                                   
#> [149] "Trichloromethane (chloroform) Total"                                               
#> [150] "Trifluralin Total"                                                                 
#> [151] "Turbidity"                                                                         
#> [152] "Uranium Dissolved"                                                                 
#> [153] "Uranium Total"                                                                     
#> [154] "Xylene Total"                                                                      
#> [155] "Zinc Dissolved"                                                                    
#> [156] "Zinc Total"                                                                        

lookup_variables(c("AL-D", "EMS_AS_T", "B--T", "KRYP"), messages = TRUE)

#> Replaced 'AL-D' with 'Aluminum Dissolved', 'AS-T' with 'Arsenic Total' and 'B--T' with 'Boron Total'.

#> Failed to replace 'KRYP'.

#> [1] "Aluminum Dissolved" "Arsenic Total"      "Boron Total"       
#> [4] NA